SIS4415_FinalProject_AnibalFalcon/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.13.3
strawberry-graphql[fastapi]==0.243.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
python-dateutil==2.9.0
pytz==2024.2
email-validator==2.1.0